Removed an uneeded offset and added the NOP function.
Should now compile and run. Right now hard coded to only execute 6
instructions. The next goal is to properly run a simple program to add 2
numbers together and store the result. Once this is working we can try
more complex stuff. There is also a lot of clean up that should be done
as soon as the simple tests are passing.