Fix some silly errors with the ROM file parsing code
Had a "how the hell did that even work like that" moment and fixed it.
To my relief it continued to work after my "fix" :) I was passing in the
address of a pointer into atoi....an int pointer for some reason too.
No idea why I had done that. Fortunately, I saw the compiler warnings
and fixed it.
I think early on I was thinking "woohoo, it builds! Warnings, pah!"