Corrected the opcode map and a typeo in a function name
The opcode map should now be [mostly] correct. Doing this was tricky
simply because the opcodes are not shown in one place, and the tables of
jump, destination and instruction bits were not necessarily in numeric
order

Now passes the simple test that adds 2 numbers and writes the result to
the D register.

TODO: More testing, cleanup then add timing and screen/keyboard
simulators