Add outline for dumping screen memory to file
The idea is that as an intermediate measure we will execute a program
and write its screen memory to file as we go. Then we can read this dump
file using a Python program which will display screen data.
Eventually we can have some way of getting the screen data displayed in
real time.